+++ modules/types/builtin/cpp/ischarBuiltin.cpp
@@ -7,19 +7,20 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "CallMexBuiltin.hpp"
-#include "NelsonConfiguration.hpp"
+#include "ischarBuiltin.hpp"
 #include "Error.hpp"
-#include "i18n.hpp"
+#include "OverloadFunction.hpp"
+#include "CheckerHelpers.hpp"
 //=============================================================================
-namespace Nelson {
+using namespace Nelson;
 //=============================================================================
-void
-CallMexBuiltin(void* fptr, const ArrayOfVector& inputArgs, int nargout, ArrayOfVector& outputArgs,
-    bool interleavedComplex)
+ArrayOfVector
+Nelson::TypeGateway::ischarBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
-    Error(_W("Mex not supported."));
+    ArrayOfVector retval;
+    nargoutcheck(nLhs, 0, 1);
+    nargincheck(argIn, 1, 1);
+    retval << ArrayOf::logicalConstructor(argIn[0].isCharacterArray());
+    return retval;
 }
-//=============================================================================
-} // namespace Nelson
 //=============================================================================