+++ modules/types/builtin/cpp/isint8Builtin.cpp
@@ -7,19 +7,27 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "CallMexBuiltin.hpp"
-#include "NelsonConfiguration.hpp"
+#include "isint8Builtin.hpp"
 #include "Error.hpp"
-#include "i18n.hpp"
+#include "OverloadFunction.hpp"
+#include "CheckerHelpers.hpp"
 //=============================================================================
-namespace Nelson {
+using namespace Nelson;
 //=============================================================================
-void
-CallMexBuiltin(void* fptr, const ArrayOfVector& inputArgs, int nargout, ArrayOfVector& outputArgs,
-    bool interleavedComplex)
+ArrayOfVector
+Nelson::TypeGateway::isint8Builtin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
-    Error(_W("Mex not supported."));
+    ArrayOfVector retval;
+    nargoutcheck(nLhs, 0, 1);
+    nargincheck(argIn, 1, 1);
+    bool bSuccess = false;
+    if (eval->mustOverloadBasicTypes()) {
+        retval = OverloadFunction(eval, nLhs, argIn, "isint8", bSuccess);
+    }
+    if (!bSuccess) {
+        bool bRes = (argIn[0].getDataClass() == NLS_INT8);
+        retval << ArrayOf::logicalConstructor(bRes);
+    }
+    return retval;
 }
-//=============================================================================
-} // namespace Nelson
 //=============================================================================