+++ modules/types/builtin/cpp/isstructBuiltin.cpp
@@ -7,19 +7,21 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "CallMexBuiltin.hpp"
-#include "NelsonConfiguration.hpp"
+#include "isstructBuiltin.hpp"
 #include "Error.hpp"
-#include "i18n.hpp"
+#include "CheckerHelpers.hpp"
 //=============================================================================
-namespace Nelson {
+using namespace Nelson;
 //=============================================================================
-void
-CallMexBuiltin(void* fptr, const ArrayOfVector& inputArgs, int nargout, ArrayOfVector& outputArgs,
-    bool interleavedComplex)
+ArrayOfVector
+Nelson::TypeGateway::isstructBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
-    Error(_W("Mex not supported."));
+    ArrayOfVector retval;
+    nargoutcheck(nLhs, 0, 1);
+    nargincheck(argIn, 1, 1);
+    ArrayOf paramOne = argIn[0];
+    bool bRes = paramOne.isStruct();
+    retval << ArrayOf::logicalConstructor(bRes);
+    return retval;
 }
-//=============================================================================
-} // namespace Nelson
 //=============================================================================