+++ modules/types/builtin/cpp/isuint16Builtin.cpp
@@ -7,19 +7,20 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "CallMexBuiltin.hpp"
-#include "NelsonConfiguration.hpp"
+#include "isuint16Builtin.hpp"
 #include "Error.hpp"
-#include "i18n.hpp"
+#include "CheckerHelpers.hpp"
 //=============================================================================
-namespace Nelson {
+using namespace Nelson;
 //=============================================================================
-void
-CallMexBuiltin(void* fptr, const ArrayOfVector& inputArgs, int nargout, ArrayOfVector& outputArgs,
-    bool interleavedComplex)
+ArrayOfVector
+Nelson::TypeGateway::isuint16Builtin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
-    Error(_W("Mex not supported."));
+    nargoutcheck(nLhs, 0, 1);
+    nargincheck(argIn, 1, 1);
+    ArrayOfVector retval(1);
+    bool bRes = (argIn[0].getDataClass() == NLS_UINT16);
+    retval << ArrayOf::logicalConstructor(bRes);
+    return retval;
 }
-//=============================================================================
-} // namespace Nelson
 //=============================================================================