+++ modules/types/builtin/include/isemptyBuiltin.hpp
@@ -7,19 +7,17 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "CallMexBuiltin.hpp"
-#include "NelsonConfiguration.hpp"
-#include "Error.hpp"
-#include "i18n.hpp"
+#pragma once
 //=============================================================================
-namespace Nelson {
+#include "ArrayOf.hpp"
+#include "Evaluator.hpp"
+#include "nlsTypes_builtin_exports.h"
 //=============================================================================
-void
-CallMexBuiltin(void* fptr, const ArrayOfVector& inputArgs, int nargout, ArrayOfVector& outputArgs,
-    bool interleavedComplex)
-{
-    Error(_W("Mex not supported."));
-}
+namespace Nelson::TypeGateway {
+//=============================================================================
+NLSTYPES_BUILTIN_IMPEXP
+ArrayOfVector
+isemptyBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================